pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-ai"
version = "0.1.0"
description = "Terminal LLM Assistant (Gemini) with memory, tools, and REPL"
readme = "README.md"
authors = [{ name = "JustFeo" }]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "google-generativeai>=0.7.2",
    "chromadb>=0.5.5",
    "SQLAlchemy>=2.0",
    "simpleeval>=0.9.13",
    "duckduckgo-search>=6.1.7",
    "pypdf>=4.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "black>=25.0",
    "isort>=6.0",
    "flake8>=7.0",
]

[project.scripts]
cli-ai = "cli_ai.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 100
extend-ignore = "E203, W503" 